creating a profile works, trying to get rendering profile page of user at unique url working
