+++ backend/routes/profile.ts
@@ -1,5 +1,4 @@
 import express, { RequestHandler } from "express";
-import { v4 as uuidv4 } from "uuid";
 import dynamoDB from "../db/config/dynamodb";
 import { Profile, TableNames } from "../db/schemas";
 
@@ -9,7 +8,7 @@ const router = express.Router();
 router.post("/", (async (req, res) => {
   try {
     const profile: Profile = {
-      profileId: uuidv4(),
+      profileId: req.body.profileId,
       userId: req.body.userId,
       firstName: req.body.firstName,
       lastName: req.body.lastName,
@@ -54,7 +53,7 @@ router.get("/:profileId", (async (req, res) => {
     res.json(result.Item);
   } catch (error) {
     console.error("Error fetching profile:", error);
-    res.status(500).json({ error: "Could not fetch profile" });
+    res.status(500).json({ error: `Could not fetch profile with id: ${req.params.profileId}. ${error}` });
   }
 }) as RequestHandler);
