+++ docker-compose.yml
@@ -8,8 +8,7 @@ services:
       - "5001:5001"
     environment:
       - NODE_ENV=production
-    volumes:
-    - ./backend/db/.env:/app/db/.env:ro
+    env_file: "./backend/.env"
 
   frontend:
     build:
